+++ src/Libraries/RevitNodes/GeometryConversion/RevitToProtoMesh.cs
@@ -18,7 +18,7 @@ namespace Revit.GeometryConversion
                 .Select(mesh.get_Triangle)
                 .Select(tri => IndexGroup.ByIndices(tri.get_Index(0), tri.get_Index(1), tri.get_Index(2)));
 
-            return Autodesk.DesignScript.Geometry.Mesh.ByPointsFaceIndices(pts, tris);
+            return Autodesk.DesignScript.Geometry.Mesh.ByPointsIndexGroups(pts, tris);
 
         }