+++ test/Libraries/RevitNodesTests/Elements/ParameterTests.cs
@@ -12,22 +12,6 @@ namespace RevitNodesTests.Elements
     public class ParameterTests : RevitNodeTestBase
     {
         /// <summary>
-        /// Create a Project parameter and test if it exists
-        /// </summary>
-        [Test]
-        [TestModel(@".\elementSharedParameters.rvt")]
-        public void CreateProjectParameter_ValidArgs()
-        {
-            List<Category> categories = new List<Category>() { Category.ByName("Walls") };
-
-            Parameter.CreateProjectParameter("MyParameter", "MyGroup", SpecType.FromExisting(Autodesk.Revit.DB.SpecTypeId.String.Text), GroupType.FromExisting(Autodesk.Revit.DB.GroupTypeId.Data), true, categories);
-
-            var fec = new Autodesk.Revit.DB.FilteredElementCollector(DocumentManager.Instance.CurrentDBDocument).OfClass(typeof(Autodesk.Revit.DB.Wall));
-            var wall = fec.FirstElement(); 
-            Assert.IsNotNull(wall.LookupParameter("MyParameter"));
-        }
-
-        /// <summary>
         /// Create a Shared parameter and test if it exists
         /// </summary>
         [Test]